+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Use safe `localStorage` wrapper
 - Fix `Sign Out` from the wallet widget error
 - Send only ETH account address to applications tracking endpoints
+- Fix issue with first connect in mobile browsers
 
 ### v1.22.0